Wrote the Logic for shift add
checks if the plier is zero.

rotates the LSB of the plier to the MSB and checks if the flag is set. this checks of adding is needed.

add the cand to the init_product.

shifts the cand bits to the left.
shifts the plier bits to the right.

puts the final product into r0